insertion: move enums to a dedicated file
