+++ itou/insertion/models.py
@@ -7,6 +7,7 @@ from django.db import models
 from django.db.models import BooleanField, Exists, OuterRef, Q, Value
 from django.utils import timezone
 
+from itou.insertion.enums import GenericReferenceItemKind, GenericReferenceItemSource
 from itou.utils.storage.s3 import generate_dora_storage_url
 from itou.utils.validators import validate_post_code
 
@@ -57,26 +58,6 @@ class GeolocatedAddressMixin(models.Model):
 
     class Meta:
         abstract = True
-
-
-class GenericReferenceItemSource(models.TextChoices):
-    DATA_INCLUSION = "DATA_INCLUSION", "data·inclusion"
-    DORA = "DORA", "DORA"
-
-
-class GenericReferenceItemKind(models.TextChoices):
-    FEE = "FEE", "Frais"
-    FUNDING_LABEL = "FUNDING_LABEL", "Label de financement"
-    MOBILIZATION = "MOBILIZATION", "Mode de mobilisation"
-    MOBILIZATION_BENEFICIARY = "MOBILIZATION_BENEFICIARY", "Mode de mobilisation bénéficiaires"
-    MOBILIZATION_PUBLIC = "MOBILIZATION_PUBLIC", "Personne mobilisatrices"
-    MOBILIZATION_PROFESSIONAL = "MOBILIZATION_PROFESSIONAL", "Mode de mobilisation professionnels"
-    NETWORK = "NETWORK", "Réseau porteur"
-    PUBLIC = "PUBLIC", "Public"
-    RECEPTION = "RECEPTION", "Mode d'accueil"
-    SERVICE_KIND = "SERVICE_KIND", "Type de service"
-    SOURCE = "SOURCE", "Source"
-    THEMATIC = "THEMATIC", "Thématique"
 
 
 class GenericReferenceItem(models.Model):
